backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-settings==2.1.0
motor==3.3.2

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

